+++ component-overview/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.98](https://github.com/SpareBank1/designsystem/compare/@sb1/ffe-component-index@8.4.97...@sb1/ffe-component-index@8.4.98) (2025-01-29)
+
+**Note:** Version bump only for package @sb1/ffe-component-index
+
+
+
+
+
 ## [8.4.97](https://github.com/SpareBank1/designsystem/compare/@sb1/ffe-component-index@8.4.96...@sb1/ffe-component-index@8.4.97) (2025-01-27)
 
 **Note:** Version bump only for package @sb1/ffe-component-index